feat: make another Authenticator AuthService Module
