+++ app/src/main/java/com/lyfe/android/core/data/network/di/NetworkModule.kt
@@ -2,7 +2,7 @@ package com.lyfe.android.core.data.network.di
 
 import com.lyfe.android.BuildConfig
 import com.lyfe.android.core.data.network.adapter.ResultCallAdapterFactory
-import com.lyfe.android.core.data.network.authenticator.AuthAuthenticator
+import com.lyfe.android.core.data.network.authenticator.TokenAuthenticator
 import com.lyfe.android.core.data.network.converter.asConverterFactory
 import com.lyfe.android.core.data.network.interceptor.TokenInterceptor
 import com.lyfe.android.core.data.network.service.AuthService
@@ -17,6 +17,7 @@ import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import java.util.concurrent.TimeUnit
+import javax.inject.Named
 import javax.inject.Singleton
 
 @Module
@@ -30,7 +31,7 @@ object NetworkModule {
 
 	@Provides
 	@Singleton
-	fun providesLyfeOkHttpClient(tokenInterceptor: TokenInterceptor, authAuthenticator: AuthAuthenticator): OkHttpClient =
+	fun providesLyfeOkHttpClient(tokenInterceptor: TokenInterceptor, tokenAuthenticator: TokenAuthenticator): OkHttpClient =
 		OkHttpClient.Builder()
 			.connectTimeout(ConnectTimeout, TimeUnit.SECONDS)
 			.writeTimeout(WriteTimeout, TimeUnit.SECONDS)
@@ -41,7 +42,7 @@ object NetworkModule {
 				}
 			)
 			.addInterceptor(tokenInterceptor)
-			.authenticator(authAuthenticator)
+			.authenticator(tokenAuthenticator)
 			.build()
 
 	@Provides
@@ -65,7 +66,35 @@ object NetworkModule {
 
 	@Provides
 	@Singleton
-	fun providesAuthService(retrofit: Retrofit): AuthService {
+	@Named("lyfe")
+	fun providesLyfeAuthService(retrofit: Retrofit): AuthService {
+		return retrofit.create(AuthService::class.java)
+	}
+
+	@Provides
+	@Singleton
+	@Named("authenticator")
+	fun providesAuthenticatorAuthService(): AuthService {
+		val jsonConfig = Json { isLenient = true }
+
+		val okHttpClient = OkHttpClient.Builder()
+			.connectTimeout(ConnectTimeout, TimeUnit.SECONDS)
+			.writeTimeout(WriteTimeout, TimeUnit.SECONDS)
+			.readTimeout(ReadTimeout, TimeUnit.SECONDS)
+			.addInterceptor(
+				HttpLoggingInterceptor().apply {
+					level = HttpLoggingInterceptor.Level.BODY
+				}
+			)
+			.build()
+
+		val retrofit = Retrofit.Builder()
+			.baseUrl(BuildConfig.BASE_URL)
+			.addConverterFactory(jsonConfig.asConverterFactory(contentType))
+			.addCallAdapterFactory(ResultCallAdapterFactory())
+			.client(okHttpClient)
+			.build()
+
 		return retrofit.create(AuthService::class.java)
 	}
 }
